fix(money): the registration checkout amount is int64, not a float
CheckoutResult.AmountTZS and SignupResult.AmountTZS were float64, the
only float on a money path anywhere in this fleet.

No harm has occurred and none could have at current prices, and it is
worth saying that plainly rather than overstating it. Billing produces
the value as int64 the whole way down -- plan.RegistrationCents / 100
into model.Invoice.AmountTZS int64, out through CheckoutView.AmountTZS
int64 -- and nothing on this side ever cast it back, persisted it,
summed it or sent it to a gateway. float64 holds every integer to 2^53
exactly and Go emits 2000, not 2000.0.

It is a landmine rather than a bug. The moment anyone adds arithmetic on
this field, or an int64() cast to store it, the class of bug becomes
possible -- and the type had already spread by copy into
drs-setup-service and dhs-setup, which is exactly how it propagates.

The fleet had already decided this. Libraries/defolt-kit/money declares
`type Money int64`, whole shillings, and names these float64 amount_tzs
DTO fields as the drift it retires at adoption time.
Libraries/defolt-contracts/payment already declares
SelcomPaymentEvent.AmountTZS int64. The float was the outlier against a
written house standard, not a competing convention.

Wire-compatible, so there is no deploy ordering to respect: JSON has one
number type, the field name does not change, and the emitted bytes are
identical for any integral value. Billing stores int64 and structurally
cannot emit a fraction; if it somehow did, this now fails the unmarshal
loudly instead of rounding silently, and CreateCheckout's error is
already handled as non-fatal (the tenant stays pending_payment and the
link can be re-issued).

Also bumps golang.org/x/mod to v0.40.0. CVE-2026-56864 / 56865 landed
against v0.33.0 between pipelines, and security-scan runs trivy with
--exit-code 1, so without this the push would red-line on a CVE that has
nothing to do with the change. A failed gate marks build and deploy
skipped and the old image keeps serving.

+++ service/billing_client.go
@@ -33,12 +33,28 @@ func NewBillingClient(baseURL, internalKey string) *BillingClient {
 
 // CheckoutResult mirrors the envelope data of
 // POST {billing}/api/v1/internal/tenants/{id}/checkout.
+//
+// AmountTZS is int64 WHOLE SHILLINGS. It was float64 until 2026-09-04,
+// which was the one float on a money path anywhere in this fleet.
+// Billing produces it as int64 the whole way down -- plan.RegistrationCents
+// / 100 into model.Invoice.AmountTZS int64, out through
+// service.CheckoutView.AmountTZS int64 -- and
+// Libraries/defolt-kit/money declares the fleet type as `Money int64`,
+// whole shillings, naming these float64 amount_tzs DTO fields as the
+// drift it retires. Libraries/defolt-contracts/payment already declares
+// SelcomPaymentEvent.AmountTZS int64.
+//
+// The change is wire-compatible: JSON has one number type, so the field
+// name and the emitted bytes are identical for any integral value and no
+// deploy ordering is required. Billing stores int64 and so cannot emit a
+// fraction; if it ever did, this now fails the unmarshal loudly instead
+// of rounding silently, and CreateCheckout's error is already non-fatal.
 type CheckoutResult struct {
-	InvoiceID  string  `json:"invoice_id"`
-	AmountTZS  float64 `json:"amount_tzs"`
-	PaymentURL string  `json:"payment_url"`
-	Reference  string  `json:"reference"`
-	ExpiresAt  string  `json:"expires_at"`
+	InvoiceID  string `json:"invoice_id"`
+	AmountTZS  int64  `json:"amount_tzs"`
+	PaymentURL string `json:"payment_url"`
+	Reference  string `json:"reference"`
+	ExpiresAt  string `json:"expires_at"`
 }
 
 // CreateCheckout asks billing for a Selcom checkout URL covering the
